chore: refactor ses smtp password
